% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geo}
\alias{geo}
\alias{geo.character}
\alias{geo.data.frame}
\alias{geo.list}
\alias{geo.t2d_geo}
\alias{geo.t2d_tin}
\alias{print.t2d_geo}
\title{Mesh / geometry  object}
\usage{
geo(x, fname, ...)

\method{geo}{character}(x, fname = NULL, ...)

\method{geo}{data.frame}(
  x,
  fname = NULL,
  col_x = "x",
  col_y = "y",
  col_elev = "z",
  col_elem = "id",
  title = "",
  ...
)

\method{geo}{list}(x, fname = NULL, ...)

\method{geo}{t2d_geo}(x, fname = NULL, data = NULL, ...)

\method{geo}{t2d_tin}(x, fname = NULL, ..., dem, title = "")

\method{print}{t2d_geo}(x, ...)
}
\arguments{
\item{x}{Either: A \code{character} string providing the name of an existing geometry file;
a \code{data.frame} with mesh elements (a triplet of points defines an element);
a \code{list} with all required elements (see 'Value');
an object of class \code{t2d_geo};
an object of class \code{t2d_tin} (requires input \code{dem}).}

\item{fname}{\code{character}, name for the associated geometry file (can also be used to replace an existing entry).}

\item{...}{Arguments passed to or from other methods.}

\item{col_x}{If \code{x} is a \code{data.frame}: column with the x coordinates.}

\item{col_y}{If \code{x} is a \code{data.frame}: column with the y coordinates.}

\item{col_elev}{If \code{x} is a \code{data.frame}: column with the mesh elevation.}

\item{col_elem}{If \code{x} is a \code{data.frame}: column with identifier for mesh elements (3 point are an element / triangle).}

\item{title}{\code{character}, mesh title (not necessary, max. 72 characters).}

\item{data}{\code{list} that can be given to update an existing object \code{x}.}

\item{dem}{If \code{x} is of type \code{t2d_tin}: a DEM from which the values will
be interpolated to the TIN mesh points. This input will be forwarded to
\code{\link{interpol}} (argument \code{src}; see descriptions therein for supported
types). You can pass further arguments for interpolation, e.g. the numer of
neighbours \code{n} etc.}
}
\value{
An object of class \code{t2d_geo} consisting of an attribute \code{file}
pointing to a specific geometry file and a \code{list} with the following elements:
\describe{
  \item{header}{General mesh information including title, precision (of numbers in
    the slf file in bytes), the numbers of mesh elements and points, and the number
    of points per mesh (3 in case of triangles which is the only supported value for now).}
  \item{tin}{A mesh object of class \code{t2d_tin}, see \link{tin}.}
  \item{elevation}{Values of mesh point elevations.}
}
}
\description{
Initialise a mesh (geometry) object for use within TELEMAC.
}
\note{
Also note the associated \code{\link{plot.t2d_geo}} method.

If \code{x} is a \code{data.frame}, all input arguments referring to columns
of \code{x} support \code{\link[rlang]{quasiquotation}}, i.e. they can be specified by
their names, quoted or unquoted, or as column position.
}
