% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lines.t2d_tin}
\alias{lines.t2d_tin}
\title{Add TELEMAC mesh as lines to a plot}
\usage{
\method{lines}{t2d_tin}(x, ...)
}
\arguments{
\item{x}{An object of class \code{t2d_tin}.}

\item{...}{Arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
No return value, called for side effects (plot).
}
\description{
Adds the mesh (triangles) of an object of class \code{t2d_tin}
as lines to an existing plot.
}
