% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.t2d_geo}
\alias{plot.t2d_geo}
\title{Plot TELEMAC geometry}
\usage{
\method{plot}{t2d_geo}(x, s, ...)
}
\arguments{
\item{x}{An object of class \code{t2d_geo}.}

\item{s}{\code{numeric} value defining the resolution for plotting (the mesh has to be interpolated to a grid for that).}

\item{...}{Arguments passed to \code{\link[raster]{plot}}.}
}
\description{
Plots selected objects of class \code{t2d_geo} by interpolating the underlying mesh
to a regular grid and using \code{\link[raster]{plot}} of package \code{\link[raster]{raster}}.
}
