% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.t2d_res}
\alias{plot.t2d_res}
\title{Plot TELEMAC results}
\usage{
\method{plot}{t2d_res}(x, s, v = NULL, t = 0, ...)
}
\arguments{
\item{x}{An object of class \code{t2d_res}.}

\item{s}{\code{numeric} value defining the resolution for plotting (the mesh
has to be interpolated to a grid for that).}

\item{v}{\code{character}, name of the variable that is to be plotted
(only one at a time; default is to take the first variable that can be found).}

\item{t}{\code{integer}, timestep that is to be plotted
(only one at a time; default is to plot the first timestep).}

\item{...}{Arguments passed to \code{\link[raster]{plot}}.}
}
\value{
No return value, called for side effects (plot).
}
\description{
Plots selected variable and timestep of objects of class \code{t2d_res}
by interpolating the underlying mesh to a regular grid and using \code{\link[raster]{plot}}
of package \code{\link[raster]{raster}}.
}
