% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.t2d_tin}
\alias{plot.t2d_tin}
\title{Plot TELEMAC mesh}
\usage{
\method{plot}{t2d_tin}(x, ...)
}
\arguments{
\item{x}{An object of class \code{t2d_tin}.}

\item{...}{Arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
No return value, called for side effects (plot).
}
\description{
Plots the mesh (triangles) of an object of class \code{t2d_tin}.
}
