% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steering.R
\name{read_cas}
\alias{read_cas}
\title{Read steering file (*.cas)}
\usage{
read_cas(fname)
}
\arguments{
\item{fname}{\code{character} File name of the steering file to be read.}
}
\value{
A \code{list} with steering parameters and their values.
}
\description{
Reads the steering file of a TELEMAC project.
}
\seealso{
To obtain a \code{t2d_cas} object use function \code{\link{cas}}.
}
