% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{read_geo}
\alias{read_geo}
\title{Read geometry file (*.slf)}
\usage{
read_geo(fname)
}
\arguments{
\item{fname}{\code{character} File name of the geometry file to be read.}
}
\value{
A \code{list} with header information (see output of \code{\link{read_slf_header}})
and element \code{elevation}, the mesh point elevations.
}
\description{
Reads the geometry / mesh information of a TELEMAC project from a SELAFIN file
}
