% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2d.R
\name{t2d}
\alias{t2d}
\alias{print.t2d}
\title{Initialise a TELEMAC-2D project}
\usage{
t2d(
  title = "",
  wdir = ".",
  exec = "telemac2d.py",
  cas = NULL,
  geo = NULL,
  cli = NULL,
  res = NULL
)

\method{print}{t2d}(x, ...)
}
\arguments{
\item{title}{\code{character} string giving the title of the project (argument
\code{TITLE} in the steering file).}

\item{wdir}{\code{character} string, the project directory were inputs and outputs
will be written to. TELEMAC-2D's input filenames must be relative to this directory!
Default: current working directoy.}

\item{exec}{\code{character} string specifying the TELEMAC-2D executable (system command).
Default: \code{telemac2d.py}.}

\item{cas}{Passed to \code{\link{cas}} (preferably a \code{t2d_cas}) object).}

\item{geo}{Passed to \code{\link{geo}} (preferably a \code{t2d_geo}) object).}

\item{cli}{Passed to \code{\link{cli}} (preferably a \code{t2d_cli}) object).}

\item{res}{Passed to \code{\link{results}}. Can be the name of an existing
results file or left empty at the beginning in which case an empty template
object will be generated.}

\item{x}{An object of class \code{t2d}.}

\item{...}{Optional arguments passed to \code{print} methods.}
}
\value{
An object of class \code{t2d}, that is a list with elements title, wdir,
exec, cas, geo, cli, res TODO.
}
\description{
Initialises a project for 2-dimensional hydrodynamic modelling with TELEMAC-2D.
}
\details{
First, make sure TELEMAC-2D is installed and works!
}
