% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemac.R
\docType{package}
\name{telemac-package}
\alias{telemac}
\alias{telemac-package}
\title{telemac: R Interface to the TELEMAC Model Suite}
\description{
This package includes methods for initialisation, simulation, and
visualisation of the hydrodynamic model TELEMAC-2D. To start with the package
first have a look at the vignettes: \code{browseVignettes(package = "telemac")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tpilz/telemac}
  \item Report bugs at \url{https://github.com/tpilz/telemac/issues}
}

}
\author{
\strong{Maintainer}: Tobias Pilz \email{topilz@pik-potsdam.de} (\href{https://orcid.org/0000-0002-5641-3918}{ORCID})

Other contributors:
\itemize{
  \item Potsdam Institute for Climate Impact Research (PIK) [copyright holder, funder]
}

}
\keyword{internal}
