% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{tin2grid}
\alias{tin2grid}
\alias{tin2grid.data.frame}
\alias{tin2grid.t2d_geo}
\alias{tin2grid.t2d_res}
\title{Interpolate TIN-based mesh values to grid}
\usage{
tin2grid(x, s, output, ...)

\method{tin2grid}{data.frame}(
  x,
  s,
  output = c("list", "raster"),
  ...,
  col_x = "x",
  col_y = "y",
  col_z = "z",
  tinmat
)

\method{tin2grid}{t2d_geo}(x, s, output = c("list", "raster"), ..., v = "elevation")

\method{tin2grid}{t2d_res}(x, s, output = c("list", "raster"), ..., v = NULL, t = 0)
}
\arguments{
\item{x}{Either: a \code{data.frame} with mesh coordinates and elevations;
an object of class \code{t2d_geo} or \code{t2d_res}.}

\item{s}{\code{numeric} value defining the resolution of the output grid.}

\item{output}{\code{character}, either: \code{"list"} to return a \code{list} object;
\code{"raster"} to return a \code{\link[raster]{raster}} object.}

\item{...}{Arguments passed to or from other methods.}

\item{col_x}{If \code{x} is a \code{data.frame}: column with the x coordinates.}

\item{col_y}{If \code{x} is a \code{data.frame}: column with the y coordinates.}

\item{col_z}{If \code{x} is a \code{data.frame}: column with the mesh values.}

\item{tinmat}{\code{integer} matrix of point indices (referring to x and y coordinates)
defining the mesh elements.}

\item{v}{\code{character}, name of the variable that is to be extracted and
interpolated (default is to take the first variable that can be found).
If \code{x} is of class \code{t2d_geo} \code{elevation} will be taken by default.}

\item{t}{\code{integer}, timestep that is to be extracted and interpolated
(default results in the first timestep).}
}
\value{
If \code{output == "list"}: A \code{list} with:
\describe{
  \item{x}{X coordinates of the output grid}
  \item{y}{Y coordinates of the output grid}
  \item{z}{Matrix with interpolated values, where \code{z[i,j]} points to coordinates \code{x[i]} and \code{y[j]}}
}

If \code{output == "raster"}: An object of class \code{\link[raster]{raster}}.
}
\description{
Linearly interpolates the values of a mesh based on irregular triangles to a
regular grid.
}
\note{
If \code{x} is a \code{data.frame}, all input arguments referring to columns
of \code{x} support \code{\link[rlang]{quasiquotation}}, i.e. they can be specified by
their names, quoted or unquoted, or as column position.
}
