% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{write_cli}
\alias{write_cli}
\alias{write_cli.t2d_cli}
\alias{write_cli.t2d}
\title{Write boundary file (*.cli)}
\usage{
write_cli(x)

\method{write_cli}{t2d_cli}(x)

\method{write_cli}{t2d}(x)
}
\arguments{
\item{x}{An object of class \code{t2d_cli} or \code{t2d}.}
}
\value{
Returns input \code{x} invisibly.
}
\description{
Writes boundary conditions for a TELEMAC simulation.
}
