% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{write_geo}
\alias{write_geo}
\alias{write_geo.t2d_geo}
\alias{write_geo.t2d}
\title{Write geometry file (*.slf)}
\usage{
write_geo(x)

\method{write_geo}{t2d_geo}(x)

\method{write_geo}{t2d}(x)
}
\arguments{
\item{x}{An object of class \code{t2d_geo} or \code{t2d}.}
}
\value{
Returns input \code{x} invisibly.
}
\description{
Writes geometry / mesh information into a SELAFIN file.
}
