% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionals.R
\name{write_opt}
\alias{write_opt}
\alias{write_opt.t2d_opt}
\alias{write_opt.t2d}
\title{Write optional file(s)}
\usage{
write_opt(x)

\method{write_opt}{t2d_opt}(x)

\method{write_opt}{t2d}(x)
}
\arguments{
\item{x}{An object of class \code{t2d_opt} or \code{t2d}.}
}
\value{
Returns input \code{x} invisibly.
}
\description{
Writes optional TELEMAC-2D input file(s).
}
