% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{write_results}
\alias{write_results}
\alias{write_results.t2d_res}
\alias{write_results.t2d}
\title{Write results to *.slf}
\usage{
write_results(x)

\method{write_results}{t2d_res}(x)

\method{write_results}{t2d}(x)
}
\arguments{
\item{x}{An object of class \code{t2d_res} or \code{t2d}.}
}
\value{
An updated version of input \code{x}.
}
\description{
Writes data into a SELAFIN (*.slf) file.
}
\note{
Writes only the available values (\code{x$values}) into the file and adapts
the header accordingly in case the header is referring to more variables that are
not imported.
}
