% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selafin.R
\name{write_slf_header}
\alias{write_slf_header}
\title{Write slf header}
\usage{
write_slf_header(fname, header)
}
\arguments{
\item{fname}{\code{character}, name of the SELAFIN file to be created.}

\item{header}{A \code{list} with all header information (see output of \code{\link{read_slf_header}}).}
}
\value{
No return value, called for side effects.
}
\description{
Writes the header of a SELAFIN file (*.slf).
}
\note{
If a file \code{fname} already exists, it will be silently overwritten.

Make sure the datatypes in \code{header} (\code{integer} vs. \code{double}) match
(compare with output of \code{\link{read_slf_header}}).
}
