% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selafin.R
\name{write_slf_variable}
\alias{write_slf_variable}
\title{Write slf variable}
\usage{
write_slf_variable(fname, data)
}
\arguments{
\item{fname}{\code{character}, name of the SELAFIN file (has to exist already and contain header information).}

\item{data}{A \code{list} with variable values; see output of \code{\link{read_slf_variable}}
for the required elements; in addition an element 'precision' (see \code{\link{read_slf_header}}) is needed.}
}
\value{
No return value, called for side effects.
}
\description{
Writes variable values into a SELAFIN file (*.slf).
}
