% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.t2d_geo}
\alias{plot.t2d_geo}
\title{Plot TELEMAC geometry}
\usage{
\method{plot}{t2d_geo}(x, s, v = "elevation", ...)
}
\arguments{
\item{x}{An object of class \code{t2d_geo}.}

\item{s}{\code{numeric} value defining the resolution for plotting (the mesh has to be interpolated to a grid for that).}

\item{v}{\code{character}, name of the variable that is to be plotted
(only one at a time; default is \code{elevation}).}

\item{...}{Arguments passed to \code{\link[raster]{plot}}.}
}
\value{
No return value, called for side effects (plot).
}
\description{
Plots selected objects of class \code{t2d_geo} by interpolating the underlying mesh
to a regular grid and using \code{\link[raster]{plot}} of package \code{\link[raster]{raster}}.
}
